added deploy api call
fixes #1052

started split up of test_agent_server
