+++ tests/test_app.py
@@ -21,8 +21,9 @@ import os
 import subprocess
 import pytest
 import re
-import conftest
 from threading import Timer
+
+import utils
 from inmanta import const
 import signal
 from subprocess import TimeoutExpired
@@ -36,7 +37,7 @@ def get_command(tmp_dir, stdout_log_level=None, log_file=None, log_level_log_fil
         os.mkdir(directory)
     config_file = os.path.join(root_dir, "inmanta.cfg")
 
-    port = conftest.get_free_tcp_port()
+    port = utils.get_free_tcp_port()
 
     with open(config_file, 'w+') as f:
         f.write("[config]\n")
